+++ scripts/build_datasets.sh
@@ -19,17 +19,9 @@ source "$ACTIVATOR"
 mkdir -p "$WORK_DIR" "$OUT_DIR"/{kwdlc,fuman}
 git clone git@github.com:ku-nlp/KWDLC.git "$WORK_DIR"/KWDLC
 git clone git@github.com:ku-nlp/AnnotatedFKCCorpus.git "$WORK_DIR"/AnnotatedFKCCorpus
-python "$SCRIPTS"/add_features_to_raw_corpus.py "$WORK_DIR"/KWDLC/knp "$WORK_DIR"/kwdlc/knp -j "$JOBS"
-python "$SCRIPTS"/add_features_to_raw_corpus.py "$WORK_DIR"/AnnotatedFKCCorpus/knp "$WORK_DIR"/fuman/knp -j "$JOBS"
-kyoto idsplit \
-  --corpus-dir "$WORK_DIR"/kwdlc/knp \
-  --output-dir "$OUT_DIR"/kwdlc \
-  --train "$WORK_DIR"/KWDLC/id/split_for_pas/train.id \
-  --valid "$WORK_DIR"/KWDLC/id/split_for_pas/dev.id \
-  --test "$WORK_DIR"/KWDLC/id/split_for_pas/test.id
-kyoto idsplit \
-  --corpus-dir "$WORK_DIR"/fuman/knp \
-  --output-dir "$OUT_DIR"/fuman \
-  --train "$WORK_DIR"/AnnotatedFKCCorpus/id/split_for_pas/train.id \
-  --valid "$WORK_DIR"/AnnotatedFKCCorpus/id/split_for_pas/dev.id \
-  --test "$WORK_DIR"/AnnotatedFKCCorpus/id/split_for_pas/test.id
+python "$SCRIPTS"/build_dataset.py "$WORK_DIR"/KWDLC/knp "$OUT_DIR"/kwdlc \
+  --id "$WORK_DIR"/KWDLC/id/split_for_pas \
+  -j "$JOBS"
+python "$SCRIPTS"/build_dataset.py "$WORK_DIR"/AnnotatedFKCCorpus/knp "$OUT_DIR"/fuman \
+  --id "$WORK_DIR"/AnnotatedFKCCorpus/id/split_for_pas \
+  -j "$JOBS"