+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fix scripts for building datasets to support latest rhoknp and remove dependency on kyoto-reader.
 
 ## [1.2.0] - 2022-10-27
 ### Added